requirements.txt
faust-streaming==0.8.2
kafka-python==2.0.2
jsonpickle==2.1.0
beautifulsoup4==4.9.3
selenium==4.1.0
undetected-chromedriver==3.1.0
requests
certifi
confluent-kafka[avro,json,protobuf]>=1.4.2